+++ gameEngine/templates/index.html
@@ -142,12 +142,12 @@
     <h4>This is a VERY primitive save/load system that works off of unique user IDs. For the 
       purposes of this program, enter the username "devImpersonator" for playtesting and debugging
     </h4>
+    <h4>Server Messages show here: <span id="apiStatus"></span> </h4>
     <input type="text" id="username">
-    <button id="saveNewUser" class="APIUI" onclick="saveNewUser()" >Save New File</button>
-    <button id="saveExistingUser" 
-      class="APIUI" onclick="saveExistingUser()" >Save Existing File</button>
-    <button id="loadFile" class="APIUI" onclick="load()" >Load File</button>
-    <button id="DeleteFile" class="APIUI" onclick="deleteSave()" >Delete File :(</button>
+    <button id="saveNewUser" class="APIUI">Save New File</button>
+    <button id="saveExistingUser" class="APIUI">Save Existing File</button>
+    <button id="loadFile" class="APIUI">Load File</button>
+    <button id="deleteFile" class="APIUI">Delete File :(</button>
   </div>
 </div>
